web_app/pyproject.toml
[project]
name = "TeleWinBackend"
version = "0.1.0"
description = ""
authors = [{name ="Ahmed Mohamed"}]
requires-python =">=3.13,<4.0"
#readme = "README.md"
dependencies = [
    "django>=5.2",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "requests>=2.32.3",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.8.1",
    "drf-yasg>=1.21.10",
    "psycopg2[binary]==2.9.10",
    "django-cors-headers>=4.7.0",
    "pillow>=11.2.1",
    "celery (>=5.5.2,<6.0.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath=["core/"]
